Atoms

Constructors and types

Xtal.AtomPosition — Type

AtomPosition{D} <: AbstractRealSpaceData{D}

Contains information about an atom in a crystal, including atomic number, the name of an atom, and its position. Atomic names may be set arbitrarily, but will default to the atomic symbol if not provided explicitly.

An atomic number of 0 indicates a dummy atom - this does not represent a real atom, but a position of note within the structure. If no name is given for a dummy atom, it will be given the empty string.

Xtal.AtomList — Type

AtomList{D} <: AbstractRealSpaceData{D}

A list of atomic positions with an associated basis. Atomic coordinates are given in terms of the specified basis. If the basis is a zero matrix, the coordinates are assumed to be given in angstroms.

Methods

Xtal.atomname — Function

atomname(a::AtomPosition) -> String

Returns the name of the atom. Note that this returns the name that was provided during construction, which is not guaranteed to be the same as the name of the element that has the given atomic number.

Xtal.atomicno — Function

atomicno(a::AtomPosition) -> Int

Gets the atomic number of an atom in an atomic position.

Xtal.coord — Function

coord(a::AtomPosition{D}) -> SVector{D,Float64}

Returns the coordinate associated with an atomic position.

coord(g::RealSpaceDataGrid, ind...) -> SVector{D,Float64}

Returns the Cartesian coordinate associated with a grid datum at a given index.

Xtal.natom — Function

natom(l::AtomList) -> Int

Gets the number of atoms in an AtomList.

natom(xtal::Crystal) -> Int

Returns the number of atoms in a crystal's unit cell.

This function is currently not aware of space groups or settings, so if the generating set does not contain all the atoms in the cell, it will return the wrong value.

Xtal.basis — Function

basis(g::RealSpaceDataGrid{D,T}) -> RealBasis{D}

Gets the basis vectors of a RealSpaceDataGrid.

basis(hkl::HKLData)

Returns the real-space basis associated with an HKLData object.

Xtal.cartesian — Function

cartesian(l::AtomList{D}) -> AtomList{D}

Converts an AtomList from reduced coordinates (relative to some crystal basis) to Cartesian coordinates in space.

cartesian(b::AbstractMatrix{<:Real}, a::AtomPosition{D}) -> AtomPosition{D}

Converts an AtomPosition defined in terms of basis b to a new AtomPosition defined in Cartesian coordinates.

Xtal.reduce_coords — Function

reduce_coords(basis::AbstractMatrix{<:Real}, a::AtomPosition; incell=false)

Convert a coordinate from a Cartesian basis to the crystal basis. If incell is true, the position vector components will be truncated so they lie within the cell bounds (between 0 and 1).

reduce_coords(basis::AbstractMatrix{<:Real}, a::AbstractVector{AtomPosition}; incell=false)

Convert a vector of atomic coordinates from a Cartesian basis to the crystal basis. If incell is true, the position vector components will be truncated so they lie within the cell bounds (between 0 and 1).