Crystal data

General

Electrum.lengths — Method

lengths(b::AbstractBasis) -> Vector{Float64}

Returns the lengths of the basis vectors.

lengths(x) -> Float64

Calculate the lengths of the basis vectors associated with x. These will be real space lengths (assumed to be in angstroms) for real space data, and reciprocal space lengths for reciprocal space data.

Electrum.volume — Method

volume(x) -> Float64

Calculate the volume of the basis associated with x. This will be a real space volume (assumed to be in cubic angstroms) for real space data, and reciprocal space lengths for reciprocal space data.

Electrum.basis — Method

basis(x) -> AbstractBasis

Returns the basis associated with some data in x. By default, it's assumed to be accessible via the property :basis.

The return value might be a RealBasis or a ReciprocalBasis, depending on the space in which data is represented. Use RealBasis(g) or ReciprocalBasis(g) if a specific type is needed.

Basic data grids

Electrum.AbstractDataGrid — Type

AbstractDataGrid{D,T}

Supertype for crystal data associated with a grid of entries of type T in real or reciprocal space of dimension D.

Electrum.RealSpaceDataGrid — Type

RealSpaceDataGrid{D,T} <: AbstractDataGrid{D,T}

A data grid defined in real space, containing data of type T.

By convention, indexing of RealSpaceDataGrid is zero-based. This convention is used so that datasets where the first entry corresponds to data at the origin can be indexed with zeros. However, getindex() is implemented such that the dataset may be indexed by any integer, with modulo math used to convert to an index within the grid.

Electrum.voxelsize — Function

voxelsize(g::RealSpaceDataGrid) -> Float64

Gets the size of a single voxel of a RealSpaceDataGrid.

By default, units are assumed to be cubic angstroms.

voxelsize(g::HKLData)

Gets the size of a voxel asssociated with the RealSpaceDataGrid that would be generated by performing an inverse Fourier transform on the HKLData.

Electrum.integrate — Function

integrate(g::RealSpaceDataGrid{D,T}) -> T

Performs an integration across all voxels, returning a scalar value.

integrate(f::Function, g::RealSpaceDataGrid{D,T}) -> T

Applies the function f pointwise to the elements of a datagrid, then integrates the grid across all voxels.

Electrum.HKLData — Type

HKLData{D,T} <: AbstractDataGrid{D,T}

Stores information associated with a reciprocal space basis. Data can be accessed and modified by using the G-vectors as indices. Associated k-point data is also provided; if no k-point is supplied it is assumed to be the gamma point (zero(SVector{D,Float64}))

Internally, the data is stored such that the zero frequency components are at the first indices along that dimension. The data at G-vector [0, 0, 0] is stored in the backing array's [1, 1, 1] index, and the rest of the indices correspond to reciprocal space points using the FFT convention.

Electrum.HKLDict — Type

HKLDict{D,T}

An alternative to HKLData uses a dictionary instead of an array as a backing field.

This is a more space-efficient alternative to HKLData in the case of reciprocal space data with a large number of zero components. For wavefunction data, which is often specified to some energy cutoff that corresponds to a distance in reciprocal space, there are many zero valued elements to the array. Unspecified elements in an HKLDict are assumed to be zero.

AbstractFFTs.fft — Function

fft(g::RealSpaceDataGrid) -> HKLData

Performs a fast Fourier transform on the data in a RealSpaceDataGrid and returns an HKLData.

AbstractFFTs.ifft — Function

ifft(g::HKLData) -> RealSpaceDataGrid

Performs an inverse fast Fourier transform on an HKLData and returns a RealSpaceDataGrid.

More complex data grids

Electrum.PlanewaveWavefunction — Type

PlanewaveWavefunction{D,T} <: AbstractDataGrid{D,T}

Stores the components of a wavefunction constructed from a planewave basis. Usually, the coefficient data type T will be a ComplexF32, as in DFT calculations, double precision convergence of the density will correspond to single-precision converegnce of the wavefunction.

Internally, coefficients are stored in an Array{4,T}. Indexing is then manually implemented, with a D-dimensional CartesianIndex used for accessing each coefficient associated with a G-vector. PlanewaveWavefunction instances are mutable, with getindex() and setindex!() defined for them, but they are not resizable, and the backing array should not be resized.

Electrum.PlanewaveIndex — Type

Electrum.PlanewaveIndex{D}

A special indexing type used to index the components of wavefunctions in a planewave basis.

In many computational chemistry packages, the standard indexing of wavefunction components occurs in the following canonical order: spins, k-points, bands, then the h, k, and l indices of the G-vectors associated with the coefficients. However, in many cases, users will want to select a spin, k-point, or band before selecting a G-vector index.

To keep the syntax intuitive while maintaining performance (mostly by ensuring that all of the components of a wavefunction are stored compactly) this index type ensures that the iteration occurs in a natural order for users and in an efficient order for Julia.

This also ensures that G-vectors with negative indices are handled correctly: while the canonical G-vectors are those within some defined ranges of indices, out of bounds indices are reinterpreted automatically using modulo arithmetic.

Missing docstring.

Missing docstring for Electrum.nspin(::PlanewaveWavefunction). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Electrum.nkpt(::PlanewaveWavefunction). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Electrum.nband(::PlanewaveWavefunction). Check Documenter's build log for details.

Electrum.fermi — Method

fermi(wf::PlanewaveWavefunction) -> Float64

Estimates the Fermi energy associated with a reciprocal space wavefunction using the energy and occupancy data in the PlanewaveWavefunction.

k-points

Electrum.KPoint — Type

KPoint{D} <: DenseVector{D,Float64}

Stores a k-point with an associated weight that corresponds to the number of symmetry-equivalent k-points, stored as an integer.

Electrum.KPointMesh — Type

KPointMesh{D} <: AbstractVector{KPoint{D}}

Contains a list of k-points associated with a matrix describing the mesh that was used to generate the points, and its shift off the Γ point (origin). If the mesh used to generate the points is unknown, it will be set to the zero matrix of dimension D.

A KPointMesh can be indexed as if it were an ordinary Vector{KPoint{D}}.

Band structures

Electrum.BandAtKPoint — Type

BandAtKPoint

Stores information about a band's energy and its occupancy at a specific k-point.

Electrum.BandStructure — Type

BandStructure{D}

Stores information about an electronic band structure, including the list of k-points used to generate the data (as am AbstractVector{KPoint{D}})and the band information at every k-point (as a Vector{BandAtKPoint}).

Missing docstring.

Missing docstring for Electrum.nkpt(::BandStructure). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Electrum.nband(::BandStructure). Check Documenter's build log for details.

Electrum.nband — Method

nband(b::BandAtKPoint) -> Int

Returns the number of bands associated with a k-point.

Electrum.FatBands — Type

FatBands{D}

Stores information relevant to plotting fatbands.

FatBands.bands: matrix of energies at each [kpt, band].

FatBands.projband: array of lm-decomposed band structure. [orbital, ion, band, kpt].

FatBands.cband: array of complex-valued contributions to band structure.

Density of states

Electrum.AbstractDensityOfStates — Type

AbstractDensityOfStates

Supertype for all density of states data.

Electrum.DensityOfStates — Type

DensityOfStates

Contains the total density of states information.

Electrum.ProjectedDensityOfStates — Type

ProjectedDensityOfStates

Contains projected density of states information.

Electrum.fermi — Method

fermi(d::AbstractDensityOfStates) -> Float64

Gets the Fermi energy from DOS data. There are no guarantees on the unit of energy used!

Electrum.smear — Function

smear(dos::DensityOfStates, sigma::Real)

Smears the DOS function by convoluting it with a normalized Gaussian function with width sigma.

Electrum.energies — Function

energies(d::AbstractDensityOfStates; usefermi=false) -> Vector{Float64}

Gets the range of energies in the dataset. If usefermi is set to true, the energies returned will be adjusted such that the Fermi energy is set to zero.

There are no guarantees on the unit of energy used!

Electrum.nelectrons — Function

nelectrons(d::DensityOfStates)

Gets the approximate number of electrons that are needed to reach the Fermi level.

Atomic data

Electrum.SphericalHarmonic — Type

SphericalHarmonic{Lmax}

Real spherical harmonic components up to Lmax. This can be used to describe atomic orbitals or projections of data onto atomic sites.