File types

Miscellaneous

Reading

Xtal.readXYZ — Function

readXYZ(io::IO) -> Vector{AtomPosition{3}}

Reads an XYZ file into a Vector{AtomPosition{3}}.

Xtal.readXSF — Function

readXSF3D(
    io::IO;
    spgrp::Integer = 0,
    origin::AbstractVector{<:Real} = [0, 0, 0]
    ctr::Symbol = :P
) -> CrystalWithDatasets{3}

Reads in an XCrysDen XSF file from an input stream and returns a CrystalWithDatasets{3} with all datasets that have been included within the file.

Space group and origin information are not supplied in XSF files, but they can be supplied using the spgrp and origin keyword arguments. Centering information can be provided using the ctr argument, but is overridden by a space group assignment.

readXSF3D(
    filename::AbstractString;
    spgrp::Integer = 0,
    origin::AbstractVector{<:Real} = [0, 0, 0]
    ctr::Symbol = :P
) -> CrystalWithDatasets{3}

Reads an XSF file at path filename.

Xtal.readCPcoeff — Function

readCPcoeff(io::IO, Lmax::Val{L}) -> SphericalComponents{L}

Reads in the spherical harmonic projection coefficients from a CPpackage2 calculation.

By default, CPpackage2 gives the coefficients for spherical harmonics up to a maximum l value of 6.

Xtal.readCPgeo — Function

readCPgeo(io::IO) -> Vector{AtomPosition{3}}

Reads the atomic positions used for a CPpackage2 calculation.

Xtal.readCPcell — Function

readCPcell(io::IO) -> RealBasis{3}

Reads the basis vectors of the unit cell used for a CPpackage2 calculation.

Reading

Xtal.writeXYZ — Function

writeXYZ(io::IO, data::AbstractVector{<:AtomPosition})
writeXYZ(io::IO, data::AtomList)
writeXYZ(io::IO, data::AbstractCrystal)

Write an XYZ file based on a set of atomic coordinates.

Xtal.writeXSF — Function

writeXSF(io, xtal::Crystal{D})

Writes the crystal component of an XCrysDen XSF file.

writeXSF(io::IO, key, data::RealSpaceDataGrid{D,T}; periodic=true)

Writes the crystal component of an XCrysDen XSF file. By default, automatic wrapping of the datagrid occurs (values are repeated at the end of each dimension).

writeXSF(io::IO, xtaldata::CrystalWithDatasets{D,K,V})

Writes a CrystalWithDatasets to an XCrysDen XSF file.

ABINIT files

Reading

Xtal.read_abinit_density — Function

read_abinit_density(filename::AbstractString)
    -> CrystalWithDatasets{3,String,RealSpaceDataGrid{3,Float64}}

Reads a FORTRAN binary formatted abinit density file. By default, abinit density files will have the suffix DEN, but no assumptions are made about suffixes.

The header is used to automatically determine the file format, so this should read in any abinit density output (provided a function exists to parse that header).

The number of datasets returned depends on the value of nsppol in the header, as calculations with explicit treatment of spin will return spin densities.

Depending on the value of cplex, the datagrid(s) returned may be real or complex-valued.

Xtal.read_abinit_potential — Function

read_abinit_potential(filename::AbstractString)
    -> CrystalWithDatasets{3,String,RealSpaceDataGrid{3,T}} where T<:Union{Float64,ComplexF64}

Reads a FORTRAN binary formatted abinit potential file.

By default, abinit potential files will end in POT for the external potential, VHA for the Hartree potential, VXC for the exchange-correlation potential, and VHXC for the sum of both the Hartree and exchange-correlation potentials.

The header is used to automatically determine the file format, so this should read in any abinit density output (provided a function exists to parse that header).

The number of datasets returned depends on the value of nsppol in the header, as calculations with explicit treatment of spin will return spin-dependent potentials.

Depending on the value of cplex, the datagrid(s) returned may be real or complex-valued.

Xtal.read_abinit_wavefunction — Function

read_abinit_wavefunction(filename::AbstractString)
    -> CrystalWithDatasets{3,String,ReciprocalWavefunction{3,Float64}}

Reads a FORTRAN binary formatted abinit potential file.

By default, abinit potential files will end in POT for the external potential, VHA for the Hartree potential, VXC for the exchange-correlation potential, and VHXC for the sum of both the Hartree and exchange-correlation potentials.

The header is used to automatically determine the file format, so this should read in any abinit density output (provided a function exists to parse that header).

The number of datasets returned depends on the value of nsppol in the header, as calculations with explicit treatment of spin will return spin-dependent potentials.

Depending on the value of cplex, the datagrid(s) returned may be real or complex-valued.

VASP files

Reading

Xtal.readDOSCAR — Function

readDOSCAR(io::IO) -> Tuple{DensityOfStates, Vector{ProjectedDensityOfStates}}

Reads a DOSCAR file from VASP and returns its data as a tuple containing the total and projected density of states (if present).

Xtal.readKPOINTS — Function

readKPOINTS(io::IO) -> KPointGrid{3}

Reads a KPOINTS file to get the k-point mesh. So far only supports regular mesh.

Xtal.readPOSCAR — Function

readPOSCAR(io::IO) -> Crystal{3}
readCONTCAR(io::IO) -> Crystal{3}

Reads a VASP POSCAR or CONTCAR file.

A POSCAR contains the basis vectors of the system (potentially given with a scaling factor), the positions of all atoms as either Cartesian or reduced coordinates, and potentially information needed to perform an ab initio MD run.

A CONTCAR file is written at the end of a VASP run and contains the atomic coordinates after the calculation completed. This is relevant for geometry optimizations.

Xtal.readPROCAR — Function

readPROCAR(io::IO) -> FatBands{3}

Reads an lm-decomposed PROCAR file from VASP and returns its data as a FatBands struct.

Xtal.readWAVECAR — Function

readWAVECAR(io::IO) -> ReciprocalWavefunction{3,Float32}

Reads a WAVECAR file output from a VASP 4.6 calcuation.

Information about VASP WAVECAR files and much of the code was pulled from the WaveTrans website (originally written in FORTRAN): https://www.andrew.cmu.edu/user/feenstra/wavetrans/

This function is limited to WAVECAR files which have an RTAG value of 45200 (meaning the data is given as a Complex{Float32}) and have only a collinear magnetic field applied, like WaveTrans. It should also be noted that the weights of the k-points are not present in the WAVECAR file, and are set to 1 by default.

Writing

Xtal.writePOSCAR4 — Function

writePOSCAR4(
    io::IO,
    list::AtomList;
    comment = Written by Xtal.jl,
    names = false
)
writePOSCAR4(io::IO, xtal::AbstractCrystal; kwargs...)
writePOSCAR4(filename::AbstractString, data; kwargs...)

Writes crystal data to a VASP 4.6 POSCAR output. Dummy atoms are not written.

The first line, normally used to describe the system, may be altered by passing a printable object to comment.

By default, atom names are not written (since this seems to break VASP 4.6) but this may be overridden by setting names to true.

LAMMPS

Reading

Writing

Xtal.write_lammps_data — Function

write_lammps_data(
    io::IO,
    list::AtomList{D}
    sc::AbstractVecOrMat{<:Integer} = ones(Int,D);
    dummy::Bool = false
)

Writes crystal information to a LAMMPS data format that can be used to define a simulation box for running a molecular dynamics simulation. If provided, sc can be used to generate a supercell

This function currently only works correctly for 3D systems.

write_lammps_data(
    io::IO,
    xtal::AbstractCrystal{D};
    sc = diagm(ones(Int, D))::AbstractVector{<:Integer};
    dummy::Bool = false 
)

Writes crystal information to a LAMMPS data format that can be used to define a simulation box for running a molecular dynamics simulation.

This function currently only works for 3D systems.