Traits

Space traits

Electrum.BySpace — Type

BySpace <: Electrum.DataSpace

Supertype for the ByRealSpace and ByReciprocalSpace traits.

Electrum.ByRealSpace — Type

ByRealSpace <: BySpace

Trait for real space data.

Electrum.ByReciprocalSpace — Type

ByReciprocalSpace <: BySpace

Trait for reciprocal space data.

Base.inv — Method

inv(::Type{<:BySpace}) -> Type{<:BySpace}
inv(::BySpace) -> BySpace

Returns the space trait dual to the given space trait. Either a type or an instance may be given.

Examples

julia> inv(ByReciprocalSpace)
ByRealSpace

julia> inv(ByRealSpace())
ByReciprocalSpace()

Coordinate traits

Electrum.ByCoordinate — Type

ByCoordinate <: Electrum.CrystalDataTrait

Describes the coordinate system associated with data. This includes ByCartesianCoordinate and ByFractionalCoordinate.

Electrum.ByCartesianCoordinate — Type

ByCartesianCoordinate <: ByCoordinate

Trait type for coordinates represented in terms of an implicit orthonormal basis with units of bohr or rad*bohr⁻¹.

Electrum.ByFractionalCoordinate — Type

ByFractionalCoordinate <: ByCoordinate

Trait type for coordinates in D dimensions whose values depend on a choice of basis, usually the basis vectors describing the lattice in which the coordinate is contained.