Geometry

The starting point for solid state structures is a periodic lattice. Electrum provides convenient types and functions for working with lattices of arbitrary dimension, not just three dimensions.

Traits

Julia parametric types can be used to encode information about data types that are wrapped by a new composite type: for instance, Julia's Array{T,D} type includes information about the types of the elements in the type parameter T. However, type parameters can be used as tags that provide additional information about data types without relying on inheritance from another abstract type. One example of this is Ptr{T}: although a pointer is underlyingly a UInt, the type parameter encodes the type of data pointed at.

Electrum uses several traits to encode information about numerical data, particularly vectors. If a user is given an instance of SVector{3,Float64}, the data type itself does not encode important information, like whether the data corresponds to a real or reciprocal space coordinate, or whether the coordinates are fractional coordinate or Cartesian coordinates. Electrum's data types use these traits in their declarations to convey this extra information.

Conventions

Trait types are declared as singleton structs (no fields). If more information needs to be encoded, use a type parameter. The names of traits generally tend to start with By.

If creating a new data type that uses one of the traits mentioned below, use the type itself as a parameter, not the singleton instance.

Real and reciprocal space

The BySpace supertype contains two types, ByRealSpace and ByReciprocalSpace. These types are used to denote whether data is associated with real space (e.g. electron density) or reciprocal space (e.g. the Fourier transform of the electron density). When working with lattices, it is important to distinguish the two types of lattice: this is the primary reason why bare SMatrix{D,D,T} instances are not used in this package.

The units associated with data bearing the Electrum.ByRealSpace trait are bohr or powers of bohr, whereas those associated with Electrum.ByReciprocalSpace are rad bohr⁻¹ or powers thereof.

Electrum.BySpace — Type

BySpace <: Electrum.DataSpace

Supertype for the ByRealSpace and ByReciprocalSpace traits.

Electrum.ByRealSpace — Type

ByRealSpace <: BySpace

Trait for real space data.

Electrum.ByReciprocalSpace — Type

ByReciprocalSpace <: BySpace

Trait for reciprocal space data.

Coordinate system

The Electrum.ByCoordinate{D} supertype contains Electrum.ByCartesianCoordinate{D} and Electrum.ByFractionalCoordinate{D}, corresponding to Cartesian or fractional coordinates in D dimensions.

Electrum.ByCoordinate — Type

ByCoordinate <: Electrum.CrystalDataTrait

Describes the coordinate system associated with data. This includes ByCartesianCoordinate and ByFractionalCoordinate.

Electrum.ByCartesianCoordinate — Type

ByCartesianCoordinate <: ByCoordinate

Trait type for coordinates represented in terms of an implicit orthonormal basis with units of bohr or rad*bohr⁻¹.

Electrum.ByFractionalCoordinate — Type

ByFractionalCoordinate <: ByCoordinate

Trait type for coordinates in D dimensions whose values depend on a choice of basis, usually the basis vectors describing the lattice in which the coordinate is contained.

Coordinate vectors

The traits above can be incorporated into new coordinate types that wrap a SVector{D,T<:Real}, and retain information about what type of information is being stored by the coordinate.

CoordinateVector

The CoordinateVector{S,C,D,T} type is the primary data type for working with spatial coordinates. The RealCartesianCoordinate, RealFractionalCoordinate, ReciprocalCartesianCoordinate, and ReciprocalFractionalCoordinate aliases correspond to the four possible combinations of S and C type parameters, and are preferred for interactive use.

This type behaves almost identically to SVector{D,T}, but forbids nonsenical operations between data types: for instance, adding a RealCartesianCoordinate to a ReciprocalCartesianCoordinate.

Electrum.CoordinateVector — Type

CoordinateVector{S<:BySpace,C<:ByCoordinate,D,T<:Real} <: StaticVector{D,T}

Represents a spatial coordinate with space given by trait S and coordinate trait C. Coordinate values must be subtypes of Real.

Electrum.RealCartesianCoordinate — Type

RealCartesianCoordinate{D,T}
(alias for CoordinateVector{Electrum.ByRealSpace,Electrum.ByCartesianCoordinate,D,T})

Represents a real space Cartesian coordinate. For more information, see CoordinateVector.

Electrum.RealFractionalCoordinate — Type

RealFractionalCoordinate{D,T}
(alias for CoordinateVector{Electrum.ByRealSpace,Electrum.ByFractionalCoordinate,D,T})

Represents a real space fractional coordinate. For more information, see CoordinateVector.

Electrum.ReciprocalCartesianCoordinate — Type

ReciprocalCartesianCoordinate{D,T}
(alias for CoordinateVector{Electrum.ByReciprocalSpace,Electrum.ByCartesianCoordinate,D,T})

Represents a reciprocal space Cartesian coordinate. For more information, see CoordinateVector.

Electrum.ReciprocalFractionalCoordinate — Type

ReciprocalFractionalCoordinate{D,T}
(alias for CoordinateVector{Electrum.ByReciprocalSpace,Electrum.ByFractionalCoordinate,D,T})

Represents a reciprocal space fractional coordinate. For more information, see CoordinateVector.

ShiftVector

A ShiftVector is almost identical to a CoordinateVector, and wraps one as a field, but includes a weight parameter that defaults to 1.

The primary use for ShiftVector is to provide information about how data associated with a lattice is shifted with respect to the origin. In particular, it forms the implementation of KPoint{D,T}, which is simply an alias for ShiftVector{ByReciprocalSpace,D,T}. In many cases, lists of k-points are symmetry-reduced, and the weight parameter is used to account for coordinates not present in the list due to symmetry reduction.

To retain the

Electrum.ShiftVector — Type

ShiftVector{S<:BySpace,D,T} <: StaticVector{D,T}

A vector in fractional coordinates representing a shift of a lattice or lattice dataset from the origin. This wraps a CoordinateVector{S,ByFractionalCoordinate,D,T} with an optional weight parameter that may be useful when working with symmetrical structures or k-points in the irreducible Brillouin zone. If it is not explicitly set, it defaults to 1.

If constructors do not explicitly reference an element type, the element type is automatically inferred by promoting the types of the arguments.

Type aliases

ShiftVector is a general way of working with vectors which shift a lattice or data within it, and for this reason we define an alias for representing k-points:

const KPoint = ShiftVector{ByReciprocalSpace}

Electrum.KPoint — Type

KPoint{D,T} (alias for ShiftVector{ByReciprocalSpace,D,T})

Represents a k-point, or an offset from the Γ point (origin) of reciprocal space, often used in band structures, wavefunctions, or other electronic data.

For more information about this type, see ShiftVector.

Electrum.weight — Function

weight(k::ShiftVector{S,D,T}) -> T

Returns the weight associated with a ShiftVector.

Lattices

LatticeBasis{S<:BySpace,D,T} <: StaticMatrix{D,D,T} represents the real or reciprocal space basis vectors of a lattice as a matrix. For convenience, we define RealBasis as an alias for LatticeBasis{ByRealSpace}, and ReciprocalBasis for LatticeBasis{ByReciprocalSpace}.

The units of RealBasis are bohr, and those of ReciprocalBasis are radians over bohr, corresponding with the convention that the dot product of a real basis vector with a corresponding reciprocal basis vector is 2π.

Construction

A RealBasis or ReciprocalBasis can be constructed from an AbstractMatrix, Tuple, or iterator of the correct size.

In the case of a StaticMatrix, the size and element type are already known, so the constructors can simply be called as RealBasis or ReciprocalBasis:

julia> RealBasis(SMatrix{3,3}(1, 0, 0, 0, 2, 0, 0, 0, 3))
Electrum.LatticeBasis{ByRealSpace, 3, Int64}:
    a: [  1.000000   0.000000   0.000000 ]   (1.000000 bohr)
    b: [  0.000000   2.000000   0.000000 ]   (2.000000 bohr)
    c: [  0.000000   0.000000   3.000000 ]   (3.000000 bohr)

However, in the case of a Matrix or other dynamically sized data, the dimension is not known and should be supplied to avoid an exception being thrown. This is to avoid type instability arising from determining the size at runtime:

julia> RealBasis{3}([1 0 0; 0 2 0; 0 0 3])
Electrum.LatticeBasis{ByRealSpace, 3, Int64}:
    a: [  1.000000   0.000000   0.000000 ]   (1.000000 bohr)
    b: [  0.000000   2.000000   0.000000 ]   (2.000000 bohr)
    c: [  0.000000   0.000000   3.000000 ]   (3.000000 bohr)

In either case, you can supply the element type (though it must be proceeded by the dimension in all cases) to convert the input elements to the desired type (here, it's Float32):

julia> RealBasis{3,Float32}(SMatrix{3,3}(1, 0, 0, 0, 2, 0, 0, 0, 3))
Electrum.LatticeBasis{ByRealSpace, 3, Float32}:
    a: [  1.000000   0.000000   0.000000 ]   (1.000000 bohr)
    b: [  0.000000   2.000000   0.000000 ]   (2.000000 bohr)
    c: [  0.000000   0.000000   3.000000 ]   (3.000000 bohr)

Conversion

A RealBasis can be converted to a ReciprocalBasis via convert or their constructors, and vice versa:

b = ReciprocalBasis(a)
c = convert(RealBasis, b)

Tip

Avoid needless back-and-forth conversions between RealBasis and ReciprocalBasis to avoid numerical instabilities.

Mathematical operations

The RealBasis and ReciprocalBasis types support the majority of common operations used in solid-state chemistry, including addition, subtraction, multiplication, left division, and right division.

Importantly, it supports the QR decomposition provided by LinearAlgebra.qr. This decomposition is useful in that it generates a $Q$ factor, which is an orthogonal matrix (representing a Euclidean point isometry - compositions of rotations and reflections) and an $R$ factor, which is an upper triangular matrix. This operation is useful in converting lattices to a standard orientation: in the case of a QR decomposition, the $R$ factor places the first basis vector of the lattice along the first basis vector of space. In 3D, this means that $\vec{a}$ is collinear with $\vec{x}$.

Calling LinearAlgebra.qr(::LatticeBasis) returns a StaticArrays.QR{SMatrix{D,D,T,D^2}, SMatrix{D,D,T,D^2}, SVector{D,Int}}, so the $Q$ and $R$ factors are bare $SMatrix$ instances. While this is fine for the $Q$ matrix, the $R$ matrix represents a basis, and we expect an LatticeBasis return value. The triangularize function returns the $R$ factor of a QR decomposition as an LatticeBasis, discarding the $Q$ factor.

Electrum.triangularize — Function

triangularize(l::T) where T<:LatticeBasis -> T

Converts a set of basis vectors to an upper triangular form using QR decomposition.

triangularize(l::T, sc::AbstractMatrix{<:Integer}) where T<:LatticeBasis -> T

Converts a set of basis vectors to an upper triangular form using QR decomposition, with an included conversion to a larger supercell. The resulting matrix that describes the basis vectors will have only positive values along the diagonal, and therefore, is always right-handed (regardless of the transformation matrix used).

LAMMPS expects that basis vectors are given in this format.

Note

Support for the corresponding LQ decomposition (where $L$ is a lower triangular matrix) is not yet implemented.

Implementation details

The SMatrix{D1,D2,T,L} type requires four type parameters - D1 and D2 are each of the matrix dimensions, and T is the element type of the matrix. However, one last parameter is needed: L, the length of the NTuple that backs the SMatrix.

Julia currently does not allow for the calculation of type parameters from other type parameters, which poses a problem in the declaration of structs. Technically, a type like SMatrix{3,3,Float64} is an abstract type, as the L parameter is undeclared, even though the value of L is always D1 * D2. By declaring a struct to have this type, there seems to be a performance drop.

The LatticeBasis types wrap an SVector{D,SVector{D,Float64}}. This only requires a single type parameter D, and allows for fully concrete struct declarations. The :matrix property converts this data to an SMatrix{D,D,T,D^2} instance, and methods needing to access something that looks like a matrix reference this property.

Basis vectors in composite types

Some types (such as Electrum.DataGrid) store a set of basis vectors as part of the dataset. The basis function allows a user to access that set of basis vectors. By default, basis(x) returns x.basis, so defining a field basis::RealBasis{...} or basis::ReciprocalBasis{...} implements these functions automatically.

Warning

basis(x) returns an LatticeBasis, but there is no guarantee whether the return type is RealBasis or ReciprocalBasis. Use convert(::Type{<:LatticeBasis}, basis(x)) to ensure that the return type is what you expect.

The type of basis vectors stored also allows for inference of the data space trait, as the default definition of DataSpace(::Type{T}) is fieldtype(T, :basis). We encourage you to choose your basis vector type to match the data you wish to represent.

Note

For performance reasons, we encourage you to define struct fields with concrete types, and use type parameters in your struct definitions to ensure that the type is concrete.