Crystal data

General

Electrum.lengths — Method

lengths(b::LatticeBasis{S,D}) -> SVector{D}

Returns the lengths of the basis vectors. The units correspond to the type of the basis vectors: for RealBasis the units are bohr, and for ReciprocalBasis the units are rad*bohr⁻¹.

Missing docstring.

Missing docstring for Electrum.volume(::Any). Check Documenter's build log for details.

Electrum.basis — Method

basis(x)

Returns the lattice basis associated with a data structure. By default, this returns getproperty(x, :basis). This may be implemented for custom data types by either adding a method to basis() or by defining custom getproperty() and propertynames() methods.

Although basis(x) should always return an Electrum.LatticeBasis, the exact return type may vary. For predictable results, use convert(T, basis(x)) where T is the desired type.

Basic data grids

Electrum.DataGrid — Type

Electrum.DataGrid{D,B<:LatticeBasis,S<:AbstractVector{<:Real},T} <: AbstractArray{T,D}

Stores a grid of values of type T defined in a D-dimensional crystal lattice basis of type B with a shift parameter of type S.

By convention, indexing of DataGrid is zero-based. This convention is used so that the first entry corresponds to data at the origin can be indexed with zeros. However, getindex() is implemented such that the dataset may be indexed by any integer, with modulo math used to convert to an index within the grid. This is done with Electrum.reinterpret_index().

Linear indexing is one-based like that of the underlying array.

For convenience, the aliases RealDataGrid and ReciprocalDataGrid are provided and are defined below:

const RealDataGrid{D} = DataGrid{RealBasis{D,Float64},SVector{D,Float64},D}
const ReciprocalDataGrid{D} = DataGrid{ReciprocalBasis{D,Float64},KPoint{D},D}

Note that ReciprocalDataGrid uses a KPoint{D} to represent the shift, as opposed to an SVector{D}. This allows for the concurrent storage of a weight along with the shift, which may be relevant for wavefunctions which exploit the symmetry of the k-point mesh.

Missing docstring.

Missing docstring for Electrum.RealDataGrid. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Electrum.ReciprocalDataGrid. Check Documenter's build log for details.

AbstractFFTs.fft — Function

fft(g::RealDataGrid) -> ReciprocalDataGrid
fft(g::ReciprocalDataGrid) -> RealDataGrid

Performs a fast Fourier transform on the data in a DataGrid.

AbstractFFTs.ifft — Function

ifft(g::RealDataGrid) -> ReciprocalDataGrid
ifft(g::ReciprocalDataGrid) -> RealDataGrid

Performs an inverse fast Fourier transform on the data in a DataGrid.

The inverse FFT is normalized so that ifft(fft(g)) ≈ g (to within floating point error).

Electrum.voxelsize — Function

voxelsize(g::DataGrid) -> eltype(basis(g))

Gets the size of a single real space voxel associated with a data grid. Units are assumed to be bohr³.

Electrum.integrate — Function

integrate([f::Function = identity], g::RealDataGrid{D,T}) -> T

Applies the function f elementwise to a datagrid, then integrates the grid across all voxels, accounting for the voxel volumes.

Electrum.partial_derivative — Function

partial_derivative(g::RealDataGrid, dim::Integer)

Calculates the partial derivative of g along coordinates with respect to the basis vector of dimension dim.

Electrum.gradient — Function

gradient(g::RealDataGrid{D}) -> RealDataGrid{D,<:SVector{D}}

Calculates the gradient of a RealDataGrid, which consists of vectors containing the components of the partial derivatives along the orthonormal real space basis. The units of the output are those of the input multiplied by inverse bohr.

Electrum.directional_derivative — Function

directional_derivative(g::RealDataGrid{D}, v::AbstractVector) -> RealDataGrid{D}

Calculates the directional derivative of a RealDataGrid along a vector v in Cartesian coordinates. Note that v does not need to be normalized, nor is it automatically normalized; an unnormalized input will scale the results accordingly.

Missing docstring.

Missing docstring for Electrum.remove_shift. Check Documenter's build log for details.

Wavefunctions

Electrum.PlanewaveWavefunction — Type

PlanewaveWavefunction{D,T} <: AbstractArray{T,D}

Stores the components of a wavefunction constructed from a planewave basis. Usually, the coefficient data type T will be a ComplexF32, as in DFT calculations, double precision convergence of the density will correspond to single-precision converegnce of the wavefunction.

Internally, coefficients are stored in an Array{4,T}. Indexing is then manually implemented, with a D-dimensional CartesianIndex used for accessing each coefficient associated with a G-vector. PlanewaveWavefunction instances are mutable, with getindex() and setindex!() defined for them, but they are not resizable, and the backing array should not be resized.

Electrum.PlanewaveIndex — Type

Electrum.PlanewaveIndex{D}

A special indexing type used to index the components of wavefunctions in a planewave basis.

In many computational chemistry packages, the standard indexing of wavefunction components occurs in the following canonical order: spins, k-points, bands, then the h, k, and l indices of the G-vectors associated with the coefficients. However, in many cases, users will want to select a spin, k-point, or band before selecting a G-vector index.

To keep the syntax intuitive while maintaining performance (mostly by ensuring that all of the components of a wavefunction are stored compactly) this index type ensures that the iteration occurs in a natural order for users and in an efficient order for Julia.

This also ensures that G-vectors with negative indices are handled correctly: while the canonical G-vectors are those within some defined ranges of indices, out of bounds indices are reinterpreted automatically using modulo arithmetic.

Missing docstring.

Missing docstring for Electrum.nspin(::PlanewaveWavefunction). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Electrum.nkpt(::PlanewaveWavefunction). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Electrum.nband(::PlanewaveWavefunction). Check Documenter's build log for details.

Electrum.fermi — Method

fermi(wf::PlanewaveWavefunction) -> Float64

Estimates the Fermi energy associated with a reciprocal space wavefunction using the energy and occupancy data in the PlanewaveWavefunction.

k-points

Electrum.KPoint — Type

KPoint{D} <: DenseVector{D,Float64}

Stores a k-point with an associated weight that corresponds to the number of symmetry-equivalent k-points, stored as an integer.

Electrum.KPointMesh — Type

KPointMesh{D} <: AbstractVector{KPoint{D}}

Contains a list of k-points associated with a matrix describing the mesh that was used to generate the points, and its shift off the Γ point (origin). If the mesh used to generate the points is unknown, it will be set to the zero matrix of dimension D.

A KPointMesh can be indexed as if it were an ordinary Vector{KPoint{D}}.

Band structures

Electrum.BandAtKPoint — Type

BandAtKPoint

Stores information about a band's energy and its occupancy at a specific k-point.

Electrum.BandStructure — Type

BandStructure{D}

Stores information about an electronic band structure, including the list of k-points used to generate the data (as am AbstractVector{KPoint{D}})and the band information at every k-point (as a Vector{BandAtKPoint}).

Missing docstring.

Missing docstring for Electrum.nkpt(::BandStructure). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Electrum.nband(::BandStructure). Check Documenter's build log for details.

Electrum.nband — Method

nband(b::BandAtKPoint) -> Int

Returns the number of bands associated with a k-point.

Electrum.FatBands — Type

FatBands{D}

Stores information relevant to plotting fatbands.

FatBands.bands: matrix of energies at each [kpt, band].

FatBands.projband: array of lm-decomposed band structure. [orbital, ion, band, kpt].

FatBands.cband: array of complex-valued contributions to band structure.

Density of states

Electrum.AbstractDensityOfStates — Type

AbstractDensityOfStates

Supertype for all density of states data.

Electrum.DensityOfStates — Type

DensityOfStates

Contains the total density of states information.

Electrum.ProjectedDensityOfStates — Type

ProjectedDensityOfStates

Contains projected density of states information.

Electrum.fermi — Method

fermi(d::AbstractDensityOfStates) -> Float64

Gets the Fermi energy from DOS data. There are no guarantees on the unit of energy used!

Electrum.smear — Function

smear(dos::DensityOfStates, sigma::Real)

Smears the DOS function by convoluting it with a normalized Gaussian function with width sigma.

Electrum.energies — Function

energies(d::AbstractDensityOfStates; usefermi=false) -> Vector{Float64}

Gets the range of energies in the dataset. If usefermi is set to true, the energies returned will be adjusted such that the Fermi energy is set to zero.

There are no guarantees on the unit of energy used!

Electrum.nelectrons — Function

nelectrons(d::DensityOfStates)

Gets the approximate number of electrons that are needed to reach the Fermi level.

Atomic data

Electrum.SphericalHarmonic — Type

SphericalHarmonic{Lmax}

Real spherical harmonic components up to Lmax. This can be used to describe atomic orbitals or projections of data onto atomic sites.