Crystal data

General

Electrum.lengths — Method

lengths(b::AbstractBasis) -> Vector{Float64}

Returns the lengths of the basis vectors.

lengths(x) -> Float64

Calculate the lengths of the basis vectors associated with x. These will be real space lengths (assumed to be in angstroms) for real space data, and reciprocal space lengths for reciprocal space data.

Electrum.volume — Method

volume(x) -> Float64

Calculate the volume of the basis associated with x. This will be a real space volume (assumed to be in cubic angstroms) for real space data, and reciprocal space lengths for reciprocal space data.

Real space

Electrum.AbstractRealSpaceData — Type

AbstractRealSpaceData{D}

Supertype for crystal data that is defined in real space. This includes atomic coordinates.

Any basis that is stored with a subtype of this type will be a RealBasis, and calling basis() on that subtype will always return a RealBasis.

Electrum.RealSpaceDataGrid — Type

RealSpaceDataGrid{D,T} <: AbstractRealSpaceData{D}

A data grid defined in real space, containing data of type T.

By convention, indexing of RealSpaceDataGrid is zero-based. This convention is used so that datasets where the first entry corresponds to data at the origin can be indexed with zeros. However, getindex() is implemented such that the dataset may be indexed by any integer, with modulo math used to convert to an index within the grid.

Electrum.grid — Function

grid(g::RealSpaceDataGrid{D,T}) -> Array{T,D}

Creates as copy of the array that backs a RealSpaceDataGrid{D,T}, which is an Array{T,D}.

grid(hkl::HKLData{D,T}) -> Array{T,D}

Returns a copy of the array that contains the reciprocal space data.

Electrum.voxelsize — Function

voxelsize(g::RealSpaceDataGrid) -> Float64

Gets the size of a single voxel of a RealSpaceDataGrid.

By default, units are assumed to be cubic angstroms.

voxelsize(g::HKLData)

Gets the size of a voxel asssociated with the RealSpaceDataGrid that would be generated by performing an inverse Fourier transform on the HKLData.

Electrum.integrate — Function

integrate(g::RealSpaceDataGrid{D,T}) -> T

Performs an integration across all voxels, returning a scalar value.

integrate(f::Function, g::RealSpaceDataGrid{D,T}) -> T

Applies the function f pointwise to the elements of a datagrid, then integrates the grid across all voxels.

AbstractFFTs.fft — Function

fft(g::RealSpaceDataGrid) -> HKLData

Performs a fast Fourier transform on the data in a RealSpaceDataGrid and returns an HKLData.

Reciprocal space

Electrum.AbstractReciprocalSpaceData — Type

AbstractReciprocalSpaceData{D}

Supertype for crystal data that is defined in reciprocal space.

Any basis that is stored with a subtype of this type will be a ReciprocalBasis, and calling basis() on that subtype will always return a ReciprocalBasis.

Electrum.AbstractKPoints — Type

AbstractKPoints{D}

Supertype for sets of k-points.

Electrum.AbstractHKL — Type

AbstractHKL{D}

Supertype for crystal data stored by HKL index.

Electrum.KPointGrid — Type

KPointGrid{D} <: AbstractKPoints{D}

Contains a grid used to generate k-points during a calculation.

The grid itself is given as an SMatrix{D,D,Int}, and can be interpreted as a set of D vectors given in terms of the primitive basis. These vectors can alternatively be used to construct a supercell.

The shift of the k-point mesh off Γ is given as an SVector{D,Float64}.

Electrum.KPointList — Type

KPointList{D} <: AbstractKPoints{D}

Contains a list of k-points and their associated weights. This is useful when describing an ordered list of k-points that are not associated with a grid - for instance, the k-points used in band structure calculations. In the future, it will be possible to convert a KPointGrid to a KPointList.

The weights of k-points are used to compensate for their placement on sites with point symmetry. If no weights are provided, they are all assumed to be equal. Any explicit input of k-point weights will be normalized such that their sum is 1, following the convention used by abinit.

Electrum.BandAtKPoint — Type

BandAtKPoint

Stores information about a band's energy and its occupancy at a specific k-point.

Electrum.BandStructure — Type

BandStructure{D}

Stores information about an electronic band structure, including the list of k-points used to generate the data (as a KPointList{D})and the band information at every k-point (as a Vector{BandAtKPoint}).

Electrum.FatBands — Type

FatBands{D} <: AbstractReciprocalSpaceData{D}

Stores information relevant to plotting fatbands.

FatBands.bands: matrix of energies at each [kpt, band].

FatBands.projband: array of lm-decomposed band structure. [orbital, ion, band, kpt].

FatBands.cband: array of complex-valued contributions to band structure.

Electrum.HKLData — Type

HKLData{D,T} <: AbstractReciprocalSpaceData{D}

Stores information associated with a reciprocal space basis. Data can be accessed and modified by using the G-vectors as indices. Associated k-point data is also provided; if no k-point is supplied

Internally, the data is stored such that the zero frequency components are at the first indices along that dimension. The data at G-vector [0, 0, 0] is stored in the backing array's [1, 1, 1] index, and the rest of the indices correspond to reciprocal space points using the FFT convention.

Electrum.HKLDict — Type

HKLDict{D,T}

An alternative to HKLData uses a dictionary instead of an array as a backing field.

This is a more space-efficient alternative to HKLData in the case of reciprocal space data with a large number of zero components. For wavefunction data, which is often specified to some energy cutoff that corresponds to a distance in reciprocal space, there are many zero valued elements to the array. Unspecified elements in an HKLDict are assumed to be zero.

Electrum.ReciprocalWavefunction — Type

ReciprocalWavefunction{D,T<:Real} <: AbstractReciprocalSpaceData{D}

Contains a wavefunction stored by k-points and bands in a planewave basis. Used to store data in VASP WAVECAR files. Each k-point is expected to have the same number of bands.

Every band has associated data containing coefficients of the constituent planewaves stored in a HKLData{D,Complex{T}}. Unlike most data structures provided by this package, the type of complex number used does not default to Float64: wavefunction data is often supplied as a Complex{Float32} since wavefunctions usually only converge to single precision, and Float64 storage would waste space.

The energies and occupancies are also stored in fields with the corresponding names, and can be accessed by spins, k-points, and bands, with indices in that order.

Electrum.nkpt — Function

nkpt(k::AbstractKPoints{D}) -> Int

Gets the number of k-points in a KPointList or a KPointGrid.

nkpt(wf::ReciprocalWavefunction) -> Int

Returns the number of k-points associated with a ReciprocalWavefunction.

Electrum.nband — Function

nband(b::BandAtKPoint) -> Int

Returns the number of bands associated with a k-point.

nband(wf::ReciprocalWavefunction) -> Int

Returns the number of bands associated with a ReciprocalWavefunction. It is assumed that the number of bands is the same for each k-point and spin.

AbstractFFTs.ifft — Function

ifft(g::HKLData) -> RealSpaceDataGrid

Performs an inverse fast Fourier transform on an HKLData and returns a RealSpaceDataGrid.

Density of states

Electrum.AbstractDensityOfStates — Type

AbstractDensityOfStates

Supertype for all density of states data.

Electrum.DensityOfStates — Type

DensityOfStates

Contains the total density of states information.

Electrum.ProjectedDensityOfStates — Type

ProjectedDensityOfStates

Contains projected density of states information.

Missing docstring.

Missing docstring for Electrum.nkpt. Check Documenter's build log for details.

Missing docstring.

Missing docstring for Electrum.nband. Check Documenter's build log for details.

Electrum.bounds — Function

bounds(wf::ReciprocalWavefunction)

Gets the range of valid G-vectors in a ReciprocalWavefunction.

Electrum.fermi — Function

fermi(wf::ReciprocalWavefunction) -> Float64

Estimates the Fermi energy associated with a reciprocal space wavefunction using the energy and occupancy data in the ReciprocalWavefunction.

fermi(d::AbstractDensityOfStates) -> Float64

Gets the Fermi energy from DOS data. There are no guarantees on the unit of energy used!

Electrum.smear — Function

smear(dos::DensityOfStates, sigma::Real)

Smears the DOS function by convoluting it with a normalized Gaussian function with width sigma.

Electrum.energies — Function

energies(d::AbstractDensityOfStates; usefermi=false) -> Vector{Float64}

Gets the range of energies in the dataset. If usefermi is set to true, the energies returned will be adjusted such that the Fermi energy is set to zero.

There are no guarantees on the unit of energy used!

Electrum.nelectrons — Function

nelectrons(d::DensityOfStates)

Gets the approximate number of electrons that are needed to reach the Fermi level.

Atomic data

Electrum.SphericalHarmonic — Type

SphericalHarmonic{Lmax}

Real spherical harmonic components up to Lmax. This can be used to describe atomic orbitals or projections of data onto atomic sites.