Crystal data

Real space

Xtal.AbstractRealSpaceData — Type

AbstractRealSpaceData{D}

Supertype for crystal data that is defined in real space. This includes atomic coordinates.

Any basis that is stored with a subtype of this type will be a RealBasis, and calling basis() on that subtype will always return a RealBasis.

Xtal.RealSpaceDataGrid — Type

RealSpaceDataGrid{D,T} <: AbstractRealSpaceData{D}

A data grid defined in real space, containing data of type T.

By convention, indexing of RealSpaceDataGrid is zero-based. This convention is used so that datasets where the first entry corresponds to data at the origin can be indexed with zeros. However, getindex() is implemented such that the dataset may be indexed by any integer, with modulo math used to convert to an index within the grid.

Xtal.grid — Function

grid(g::RealSpaceDataGrid{D,T}) -> Array{T,D}

Gets the array that backs a RealSpaceDataGrid{D,T}, which is an Array{T,D}.

grid(hkl::HKLData{D,T}) -> Array{T,D}

Returns the array that contains the reciprocal space data. Note that this causes information about the index offset to be lost!

Xtal.volume — Function

volume(b::AbstractBasis) -> Float64

Returns the volume of a unit cell defined by a matrix. This volume does not carry the sign (negative for cells that do not follow the right hand rule).

volume(g::RealSpaceDataGrid) -> Float64

Gets the crystal volume associated with a RealSpaceDataGrid.

By default, units are assumed to be cubic angstroms.

Xtal.voxelsize — Function

voxelsize(g::RealSpaceDataGrid) -> Float64

Gets the size of a single voxel of a RealSpaceDataGrid.

By default, units are assumed to be cubic angstroms.

voxelsize(g::HKLData)

Gets the size of a voxel asssociated with the RealSpaceDataGrid that would be generated by performing an inverse Fourier transform on the HKLData.

Xtal.integrate — Function

integrate(g::RealSpaceDataGrid{D,T}) -> T

Performs an integration across all voxels, returning a scalar value.

integrate(f::Function, g::RealSpaceDataGrid{D,T}) -> T

Applies the function f pointwise to the elements of a datagrid, then integrates the grid across all voxels.

AbstractFFTs.fft — Function

fft(g::RealSpaceDataGrid) -> HKLData

Performs a fast Fourier transform on the data in a RealSpaceDataGrid and returns an HKLData.

Reciprocal space

Xtal.AbstractReciprocalSpaceData — Type

AbstractReciprocalSpaceData{D}

Supertype for crystal data that is defined in reciprocal space.

Any basis that is stored with a subtype of this type will be a ReciprocalBasis, and calling basis() on that subtype will always return a ReciprocalBasis.

Xtal.AbstractKPoints — Type

AbstractKPoints{D}

Supertype for sets of k-points.

Xtal.AbstractHKL — Type

AbstractHKL{D}

Supertype for crystal data stored by HKL index.

Xtal.KPointGrid — Type

KPointGrid{D} <: AbstractKPoints{D}

Contains a grid used to generate k-points during a calculation.

The grid itself is given as an SMatrix{D,D,Int}, and can be interpreted as a set of D vectors given in terms of the primitive basis. These vectors can alternatively be used to construct a supercell.

The shift of the k-point mesh off Γ is given as an SVector{D,Float64}.

Xtal.KPointList — Type

KPointList{D} <: AbstractKPoints{D}

Contains a list of k-points and their associated weights. This is useful when describing an ordered list of k-points that are not associated with a grid - for instance, the k-points used in band structure calculations. In the future, it will be possible to convert a KPointGrid to a KPointList.

The weights of k-points are used to compensate for their placement on sites with point symmetry. If no weights are provided, they are all assumed to be equal. Any explicit input of k-points weights will be normalized such that their sum is 1, following the convention used by abinit.

Xtal.BandAtKPoint — Type

BandAtKPoint

Stores information about a band's energy and its occupancy at a specific k-point.

Xtal.BandStructure — Type

BandStructure{D}

Stores information about an electronic band structure, including the list of k-points used to generate the data (as a KPointList{D})and the band information at every k-point (as a Vector{BandAtKPoint}).

Xtal.FatBands — Type

FatBands{D} <: AbstractReciprocalSpaceData{D}

Stores information relevant to plotting fatbands.

FatBands.bands: matrix of energies at each [kpt, band].

FatBands.projband: array of lm-decomposed band structure. [orbital, ion, band, kpt].

FatBands.cband: array of complex-valued contributions to band structure.

Xtal.HKLData — Type

HKLData{D,T} <: AbstractReciprocalSpaceData{D}

Stores information associated with specific sets of reciprocal lattice vectors. Data can be accessed and modified using regular indexing, where indices may be negative.

Internally, the data is stored such that the zero frequency components are at the first indices along that dimension.

Xtal.HKLDict — Type

HKLDict{D,T}

An alternative to HKLData uses a dictionary instead of an array as a backing field.

This is a more space-efficient alternative to HKLData in the case of reciprocal space data with a large number of zero components. For wavefunction data, which is often specified to some energy cutoff that corresponds to a distance in reciprocal space, there are many zero valued elements to the array. Unspecified elements in an HKLDict are assumed to be zero.

Xtal.ReciprocalWavefunction — Type

ReciprocalWavefunction{D,T<:Real} <: AbstractReciprocalSpaceData{D}

Contains a wavefunction stored by k-points and bands in a planewave basis. Used to store data in VASP WAVECAR files. Each k-point is expected to have the same number of bands.

Every band has associated data containing coefficients of the constituent planewaves stored in a HKLData{D,Complex{T}}. Unlike most data structures provided by this package, the type of complex number used does not default to Float64: wavefunction data is often supplied as a Complex{Float32} since wavefunctions usually only converge to single precision, and Float64 storage would waste space.

Xtal.nkpt — Function

nkpt(k::KPointList{D}) -> Int

Gets the number of k-points in a KPointList.

nkpt(wf::ReciprocalWavefunction) -> Int

Returns the number of k-points associated with a ReciprocalWavefunction.

Xtal.nband — Function

nband(b::BandAtKPoint) -> Int

Returns the number of bands associated with a k-point.

nband(wf::ReciprocalWavefunction) -> Int

Returns the number of bands associated with a ReciprocalWavefunction. It is assumed that the number of bands is the same for each k-point and spin.

Density of states

Xtal.AbstractDensityOfStates
Xtal.DensityOfStates
Xtal.ProjectedDensityOfStates
Xtal.nkpt
Xtal.nband
Xtal.bounds
Xtal.fermi
Xtal.smear
Xtal.energies
Xtal.nelectrons

Atomic data

Xtal.AtomicData — Type

AtomicData{D,T}

Data associated with individual atoms in a structure.

This is a type alias for Dict{AtomPosition{D},T}. Keys are AtomPosition entries, and the values may be of any type.

Missing docstring.

Missing docstring for Xtal.SphericalComponents. Check Documenter's build log for details.