File types

Miscellaneous

Reading

Electrum.readXYZ — Function

readXYZ(io::IO) -> AtomList{3}

Reads an XYZ file into an AtomList{3}.

Electrum.readXSF — Function

readXSF3D(
    io::IO;
    spgrp::Integer = 0,
    origin::AbstractVector{<:Real} = [0, 0, 0]
    ctr::Symbol = :P
) -> CrystalWithDatasets{3}

Reads in an XCrysDen XSF file from an input stream and returns a CrystalWithDatasets{3} with all datasets that have been included within the file.

Space group and origin information are not supplied in XSF files, but they can be supplied using the spgrp and origin keyword arguments. Centering information can be provided using the ctr argument, but is overridden by a space group assignment.

readXSF3D(
    file;
    spgrp::Integer = 0,
    origin::AbstractVector{<:Real} = [0, 0, 0]
    ctr::Symbol = :P
) -> CrystalWithDatasets{3}

Reads an XSF file at path filename.

Electrum.readCPcoeff — Function

readCPcoeff(file, Lmax::Val{L}=Val{6}()) -> SphericalComponents{L}

Reads in the spherical harmonic projection coefficients from a CPpackage2 calculation.

By default, CPpackage2 gives the coefficients for spherical harmonics up to a maximum l value of 6.

Electrum.readCPgeo — Function

readCPgeo(file) -> Vector{AtomPosition{3}}

Reads the atomic positions used for a CPpackage2 calculation.

Electrum.readCPcell — Function

readCPcell(file) -> RealBasis{3}

Reads the basis vectors of the unit cell used for a CPpackage2 calculation.

Writing

Electrum.writeXYZ — Function

writeXYZ(io::IO, data::AbstractVector{<:AbstractAtomPosition})
writeXYZ(io::IO, data::AbstractAtomList)
writeXYZ(io::IO, data::AbstractCrystal)

Write an XYZ file based on a set of atomic coordinates.

Electrum.writeXSF — Function

writeXSF(io, xtal::PeriodicAtomList{D})

Writes the crystal component of an XCrysDen XSF file.

writeXSF(io::IO, key, data::RealSpaceDataGrid{D,T}; periodic=true)

Writes the crystal component of an XCrysDen XSF file. By default, automatic wrapping of the datagrid occurs (values are repeated at the end of each dimension).

writeXSF(io::IO, xtaldata::CrystalWithDatasets{D,K,V})

Writes a CrystalWithDatasets to an XCrysDen XSF file.

ABINIT files

Reading

Electrum.read_abinit_DEN — Function

read_abinit_density(file)
    -> CrystalWithDatasets{3,String,RealSpaceDataGrid{3,Float64}}

Reads a FORTRAN binary formatted abinit density file. By default, abinit density files will have the suffix DEN, but no assumptions are made about suffixes.

The header is used to automatically determine the file format, so this should read in any abinit density output (provided a function exists to parse that header).

The number of datasets returned depends on the value of nsppol in the header, as calculations with explicit treatment of spin will return spin densities.

Depending on the value of cplex, the datagrid(s) returned may be real or complex-valued.

Electrum.read_abinit_POT — Function

read_abinit_potential(file)
    -> CrystalWithDatasets{3,String,RealSpaceDataGrid{3,T<:Number}}

Reads a FORTRAN binary formatted abinit potential file.

By default, abinit potential files will end in POT for the external potential, VHA for the Hartree potential, VXC for the exchange-correlation potential, and VHXC for the sum of both the Hartree and exchange-correlation potentials.

The header is used to automatically determine the file format, so this should read in any abinit density output (provided a function exists to parse that header).

The number of datasets returned depends on the value of nsppol in the header, as calculations with explicit treatment of spin will return spin-dependent potentials.

Depending on the value of cplex, the datagrid(s) returned may be real or complex-valued.

Electrum.read_abinit_WFK — Function

read_abinit_WFK(file)
    -> CrystalWithDatasets{3,String,PlanewaveWavefunction{3,Complex{Float64}}}

Reads a FORTRAN binary formatted abinit wavefunction file.

By default, abinit returns wavefunction data in Complex{Float64} entries, and this is maintained in the return type.

The header is used to automatically determine the file format, so this should read in any abinit density output (provided a function exists to parse that header).

By default, the function is verbose, with output printed for every k-point parsed, due to the large size of the wavefunction files. If this behavior is undesirable, the quiet keyword argument may be set to true.

VASP files

Reading

Electrum.readDOSCAR — Function

readDOSCAR(file) -> Tuple{DensityOfStates, Vector{ProjectedDensityOfStates}}

Reads a DOSCAR file from VASP and returns its data as a tuple containing the total and projected density of states (if present).

Electrum.readPOSCAR — Function

readPOSCAR(file) -> PeriodicAtomList{3}

Reads a VASP POSCAR file. A POSCAR contains the basis vectors of the system (potentially given with a scaling factor), the positions of all atoms as either Cartesian or reduced coordinates, and potentially information needed to perform an ab initio MD run (currently ignored).

By default, if the provided file path is a directory, readPOSCAR() will read from a file named POSCAR in that directory. If no path is provided, a file named POSCAR in the current working directory is read.

The similar readCONTCAR function does much the same as readPOSCAR, but defaults to the file name CONTCAR instead of POSCAR.

Electrum.readPROCAR — Function

readPROCAR(file) -> FatBands{3}

Reads an lm-decomposed PROCAR file from VASP and returns its data as a FatBands{3}.

Electrum.readWAVECAR — Function

readWAVECAR(file) -> PlanewaveWavefunction{3,Complex{Float32}}

Reads a WAVECAR file output from a VASP 4.6 calcuation to a PlanewaveWavefunction.

Information about VASP WAVECAR files and much of the code was adapted from WaveTrans (originally written in FORTRAN): https://www.andrew.cmu.edu/user/feenstra/wavetrans/

This function is limited to WAVECAR files which have an RTAG value of 45200 (meaning the data is given as a Complex{Float32}) and have only a collinear magnetic field applied, like WaveTrans. It should also be noted that the weights of the k-points are not present in the WAVECAR file, and are set to 1 by default.

By default, the function is verbose, with output printed for every k-point parsed, due to the large size of the wavefunction files. If this behavior is undesirable, the quiet keyword argument may be set to true.

Writing

Electrum.writePOSCAR — Function

writePOSCAR([file = "POSCAR"], data; names = true, dummy = false, comment)

Writes crystal data to a VASP POSCAR output. The data can be a PeriodicAtomList or an AbstractCrystal. If a directory names is given instead of a file name, the data will be written to a file named POSCAR in the provided directory. If no filename is provided, it is written to POSCAR in the current directory.

By default, atom names are written, but this is known to break VASP 4.6. This may be overridden by setting names to false, but this is known to cause its own incompatibility issues: it is known that VESTA will crash if the line containing the atomic names is missing.

Dummy atoms are not are not written by default, but they may be written by setting dummy=true.

The first line, normally used to describe the system, may be altered by passing a printable object to comment.

LAMMPS

Reading

Electrum.read_lammps_data — Function

read_lammps_data(io::IO; atoms::AbstractVector = NamedAtom[]) -> PeriodicAtomList{3}

Reads a LAMMPS data file containing atomic coordinates. Currently, this function only supports 3D crystal data.

If set, the atom_types keyword can be used to define the types of atoms being read in. This can be and AbstractVector{NamedAtom}, AbstractVector{<:AbstractString}, or AbstractVector{<:Integer}. If it is not set or empty, then all atoms will be read in as dummy atoms.

Writing

Electrum.write_lammps_data — Function

write_lammps_data(io::IO, list::PeriodicAtomList, [transform]; dummy::Bool = false)

Writes crystal information to a LAMMPS data format that can be used to define a simulation box for running a molecular dynamics simulation.

If transform is supplied, the list of atoms will be converted to a supercell with the associated transformation (either a matrix, vector, or scalar). The dummy keyword determines whether dummy atoms are included in the output (false by default).

This function currently only works for 3D systems, and outputs data for the LAMMPS metal unit system - most importantly, lengths are in angstrom.

write_lammps_data(io::IO, xtal::AbstractCrystal, [transform]; dummy::Bool = false)

Writes crystal information to a LAMMPS data format that can be used to define a simulation box for for running a molecular dynamics simulation.

The list of atoms that is written is given by converting xtal to a PeriodicAtomList, which uses the supplied transformation matrix to generate all atomic positions. If transform is supplied, the transformation will be applied to the PeriodicAtomList - it does not replace the transform provided with xtal.

This function currently only works for 3D systems.

write_lammps_data(file, data, [transform]; dummy::Bool = false)

Writes a LAMMPS data file to the path given by filename.