Lattices

Constructors and types

Electrum.RealBasis — Type

RealBasis{D,T}

Represents a lattice defined in D-dimensional real space, in units of bohr.

For further information, see Electrum.LatticeBasis.

Electrum.ReciprocalBasis — Type

ReciprocalBasis{D,T}

Represents a lattice defined in D-dimensional reciprocal space, in units of rad*bohr⁻¹.

For further information, see Electrum.LatticeBasis.

Methods

Missing docstring.

Missing docstring for Electrum.lengths(). Check Documenter's build log for details.

Missing docstring.

Missing docstring for Electrum.volume(). Check Documenter's build log for details.

Electrum.angles_cos — Function

angles_cos(b::LatticeBasis{S,D}) -> SVector{binomial(D,2)}

Generates the cosines of the unit cell angles.

The angles are generated in the correct order [α, β, γ] for 3-dimensional cells. This is achieved by reversing the output of Electrum.generate_pairs(). For crystals with more spatial dimensions, this may lead to unexpected results.

Electrum.angles_rad — Function

angles_rad(b::LatticeBasis{S,D}) -> SVector{binomial(D,2)}

Returns the angles (in radians) between each pair of basis vectors.

The angles are generated in the correct order [α, β, γ] for 3-dimensional cells. This is achieved by reversing the output of Electrum.generate_pairs(). For crystals with more spatial dimensions, this may lead to unexpected results.

Electrum.angles_deg — Function

angles_deg(b::LatticeBasis{S,D}) -> SVector{binomial(D,2)}

Returns the angles (in degrees) between each pair of basis vectors.

The angles are generated in the correct order [α, β, γ] for 3-dimensional cells. This is achieved by reversing the output of Electrum.generate_pairs(). For crystals with more spatial dimensions, this may lead to unexpected results.

Electrum.maxHKLindex — Function

Electrum.maxHKLindex(b::LatticeBasis, ecut::Real; prim=true, c = 2)

Determines the maximum integer values of the reciprocal lattice vectors needed to store data out to a specific energy cutoff for a 3D lattice.

By default, the energy cutoff is assumed to be in units of Hartree, the reciprocal lattice vector lengths are assumed to be in radbohr⁻¹, and the value of c is that of the constant (2mₑ/ħ²). In Hartree atomic units, this value is 2 - but for VASP WAVECAR outputs, the value is given in eV⁻¹angstrom⁻² - see Electrum.CVASP for more information.

The functionality implemented here was taken from WaveTrans: https://www.andrew.cmu.edu/user/feenstra/wavetrans/