Lattices

Constructors and types

Xtal.AbstractBasis — Type

AbstractBasis{D}

Supertype for sets of basis vectors in D dimensions.

This supertype includes the RealBasis{D} and ReciprocalBasis{D} types, which explicitly indicate their units (assumed to be either angstroms or inverse angstroms).

Members of AbstractBasis must implement the following checks:

That the basis vectors are linearly independent and form a right-handed coordinate system,

unless an explicit zero basis is constructed (implying no periodicity).

Xtal.RealBasis — Type

RealBasis{D} <: AbstractBasis{D}

A set of real space basis vectors, assumed to be in angstroms.

Xtal.ReciprocalBasis — Type

ReciprocalBasis{D} <: AbstractBasis{D}

A set of reciprocal space basis vectors, assumed to be in inverse angstroms.

Xtal.lattice2D — Function

lattice2D(a::Real, b::Real, γ::Real) -> RealBasis{2}

Constructs a set of basis vectors in an SMatrix that correspond to a unit cell in 2D with the same length and angle parameters (in degrees).

By default, the b-vector is oriented along y. This selection corresponds to the default orientation chosen by lattice3D().

Xtal.lattice3D — Function

lattice3D(a::Real, b::Real, c::Real, α::Real, β::Real, γ::Real) -> RealBasis{3}

Constructs a set of basis vectors in an SMatrix that correspond to a unit cell in 3D with the same length and angle parameters (in degrees).

By default, the b-vector is oriented along y, and the a-vector is chosen to be perpendicular to z, leaving the c-vector to freely vary. This selection allows for the most convenient orientation of symmetry operations.

Methods

Xtal.lengths — Function

lengths(b::AbstractBasis) -> Vector{Float64}

Returns the lengths of the basis vectors.

Missing docstring.

Missing docstring for Xtal.volume. Check Documenter's build log for details.

Xtal.angles_cos — Function

angles_cos(b::AbstractBasis) -> Vector{Float64}

Generates the cosines of the unit cell angles.

The angles are generated in the correct order [α, β, γ] for 3-dimensional cells. This is achieved by reversing the output of Xtal.generate_pairs(). For crystals with more spatial dimensions, this may lead to unexpected results.

Xtal.angles_rad — Function

angles_rad(b) -> Vector{Float64}

Returns the angles (in radians) between each pair of basis vectors.

The angles are generated in the correct order [α, β, γ] for 3-dimensional cells. This is achieved by reversing the output of Xtal.generate_pairs(). For crystals with more spatial dimensions, this may lead to unexpected results.

Xtal.angles_deg — Function

angles_deg(b) -> Vector{Float64}

Returns the angles (in degrees) between each pair of basis vectors.

The angles are generated in the correct order [α, β, γ] for 3-dimensional cells. This is achieved by reversing the output of Xtal.generate_pairs(). For crystals with more spatial dimensions, this may lead to unexpected results.

Missing docstring.

Missing docstring for Xtal.maxHKLindex. Check Documenter's build log for details.